requirements.txt
tqdm
requests
baidu-aip